Move to storable container mixin
