+++ synapseclient/models/mixins/__init__.py
@@ -1,5 +1,7 @@
 from synapseclient.models.mixins.access_control import AccessControllable
+from synapseclient.models.mixins.storable_container import StorableContainer
 
 __all__ = [
     "AccessControllable",
+    "StorableContainer",
 ]
